Convex Combination of copulas.

ConvexCombCopula class

ConvexCombCopula(copulas, alpha = rep(1, length(copulas)))

Arguments

copulas

a list of copulas of same dimension

alpha

a vector of (positive) weights

Value

An instance of the ConvexCombCopula S4 class. The object represent the copula that results from a convex combinaison of other copulas, and can be used through several methods to query classical (r/d/p/v)Copula methods, etc.

Details

The ConvexCombcopula class is used to build convex combinations of copulas, with given positives weights. The rCopula and pCopula functions works for those copulas, assuming they work for the given copulas that we combined in a convex way.

See the corresponding vignette for more details about the implementation.

Examples

dataset <- apply(LifeCycleSavings,2,rank)/(nrow(LifeCycleSavings)+1)
copulas <- list(
  cbCopula(dataset[,2:3],m=10),
  cbCopula(dataset[,2:3],m=5)
)
alpha <- c(1,4)
(cop <- ConvexCombCopula(copulas,alpha))
#> This is a ConvexCombCopula , with : 
#>    dim = 2 
#>    number of copulas = 2 
#>    alpha = 0.2 0.8 
#> sub-copulas can be accessed trhough the @copulas slot