The Cort estimator

Cort class

Cort(
  x,
  p_value_for_dim_red = 0.75,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose_lvl = 1,
  slsqp_options = NULL,
  osqp_options = NULL,
  N = 999,
  force_grid = FALSE
)

Arguments

x
The data, must be provided as a matrix with each row as an observation.

p_value_for_dim_red
a p_value for the localised dimension reduction test

min_node_size
The minimum number of observation avaliable in a leaf to initialise a split.

pseudo_data
set to True if you are already providing data on the copula space.

number_max_dim
The maximum number of dimension a split occurs in. Defaults to be all of the dimensions.

verbose_lvl
numeric. set the verbosity. 0 for no ouptut and bigger you set it the most output you get.

slsqp_options
options for nloptr::slsqp to find breakpoints : you can change defaults.

osqp_options
options for the weights optimisation. You can pass a call to osqp::osqpSettings, or NULL for defaults.

N
The number of bootstrap resamples for p_values computations.

force_grid
boolean. set to TRUE to force breakpoint to be on the n-checkerboard grid.

Value

a Cort object that can be fitted easily to produce a copula estimate.

Details

This class implements the CORT algorithm to a fit a multivariate copula using piece constant density. See O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020) for the details of this density estimation procedure.

References

Laverny O, Maume-Deschamps V, Masiello E, Rullière D (2020). “Dependence Structure Estimation Using Copula Recursive Trees.” arXiv preprint arXiv:2005.02912.

Examples

(Cort(LifeCycleSavings[,1:3]))

#> Splitting...
#> 
#>      1 leaves to split...
#>      5 leaves to split...
#>      10 leaves to split...
#>      1 leaves to split...
#> Enforcing constraints...
#> Done !

#> Cort copula model: 50x3-dataset and 91 leaves.