cbCopula contructor

The cbCopula class computes a checkerboard copula with a given checkerboard parameter \(m\), as described by A. Cuberos, E. Masiello and V. Maume-Deschamps (2019). Assymptotics for this model are given by C. Genest, J. Neslehova and R. bruno (2017). The construction of this copula model is as follows :

cbCopula(x, m = rep(nrow(x), ncol(x)), pseudo = FALSE)

Arguments

x
the data to be used

m
checkerboard parameters

pseudo
Boolean, defaults to FALSE. Set to TRUE if you are already providing pseudo data into the x argument.

Value

a cbCopula object

Details

Start from a dataset with \(n\) i.i.d observation of a \(d\)-dimensional copula (or pseudo-observations), and a checkerboard parameter \(m\),dividing \(n\).

Consider the ensemble of multi-indexes \(\mathcal{I} = \{\mathbf{i} = (i_1,..,i_d) \subset \{1,...,m \}^d\}\) which indexes the boxes :

$$B_{\mathbf{i}} = \left]\frac{\mathbf{i}-1}{m},\frac{\mathbf{i}}{m}\right]$$

Let now \(\lambda\) be the dimension-unspecific lebesgue measure on any power of \(\mathbb{R}\), that is :

$$\forall d \in \mathbb{N}, \forall x,y \in \mathbb{R}^p, \lambda(\left(x,y\right)) = \prod\limits_{p=1}^{d} (y_i - x_i)$$

Let furthermore \(\mu\) and \(\hat{\mu}\) be respectively the true copula measure of the sample at hand and the classical Deheuvels empirical copula, that is :

For \(n\) i.i.d observation of the copula of dimension \(d\), let \(\forall i \in \{1,...,d\}, \, R_i^1,...,R_i^d\) be the marginal ranks for the variable \(i\).

\(\forall x \in \mathcal{I}^d, \text{ let } \hat{\mu}((0,x)) = \frac{1}{n} \sum\limits_{k=1}^n \mathbb{1}_{R_1^k\le x_1,...,R_d^k\le x_d}\)

The checkerboard copula, \(C\), and the empirical checkerboard copula, \(\hat{C}\), are then defined by the following :

$$\forall x \in (0,1)^d, C(x) = \sum\limits_{\mathbf{i}\in\mathcal{I}} {m^d \mu(B_{\mathbf{i}}) \lambda((0,\mathbf{x})\cap B_{\mathbf{i}})}$$

Where \(m^d = \lambda(B_{\mathbf{i}})\).

This copula is a special form of patchwork copulas, see F. Durante, J. Fernández Sánchez and C. Sempi (2013) and F. Durante, J. Fernández Sánchez, J. Quesada-Molina and M. Ubeda-Flores (2015). The estimator has the good property of always being a copula.

The checkerboard copula is a kind of patchwork copula that only uses independent copula as fill-in, only where there are values on the empirical data provided. To create such a copula, you should provide data and checkerboard parameters (depending on the dimension of the data).

References

Cuberos A, Masiello E, Maume-Deschamps V (2019). “Copulas Checker-Type Approximations: Application to Quantiles Estimation of Sums of Dependent Random Variables.” Communications in Statistics - Theory and Methods, 1--19. ISSN 0361-0926, 1532-415X.

Genest C, Nešlehová JG, Rémillard B (2017). “Asymptotic Behavior of the Empirical Multilinear Copula Process under Broad Conditions.” Journal of Multivariate Analysis, 159, 82--110. ISSN 0047259X.

Durante F, Fernández Sánchez J, Sempi C (2013). “Multivariate Patchwork Copulas: A Unified Approach with Applications to Partial Comonotonicity.” Insurance: Mathematics and Economics, 53(3), 897--905. ISSN 01676687.

Durante F, Fernández-Sánchez J, Quesada-Molina JJ, Úbeda-Flores M (2015). “Convergence Results for Patchwork Copulas.” European Journal of Operational Research, 247(2), 525--531. ISSN 03772217.