Copula density

This function returns the density of a given copula on given observations.

dCopula(u, copula, ...)

# S4 method for matrix,Cort
dCopula(u, copula)

# S4 method for matrix,CortForest
dCopula(u, copula)

# S4 method for matrix,cbCopula
dCopula(u, copula)

Arguments

u
numeric matrix : one row per observation

copula
the copula object

...
other parameter to be passed to methods for this generic.

Value

the density of the copula on each observation

Functions

dCopula,matrix,Cort-method: Method for the class Cort

dCopula,matrix,CortForest-method: Method for the class CortForest

dCopula,matrix,cbCopula-method: Method for the cbCopula

Examples

cop <- cbCopula(LifeCycleSavings,m = 5)
dCopula(rep(0,5),cop)

#> [1] 0.02

dCopula(rep(0.5,5),cop)

#> [1] 0.02

dCopula(rep(1,5),cop)

#> [1] 0.02