Copula cdf

This function returns the value of the copula itself on given points.

pCopula(u, copula, ...)

# S4 method for matrix,ConvexCombCopula
pCopula(u, copula)

# S4 method for matrix,Cort
pCopula(u, copula)

# S4 method for matrix,CortForest
pCopula(u, copula)

# S4 method for matrix,cbCopula
pCopula(u, copula)

# S4 method for matrix,cbkmCopula
pCopula(u, copula)

Arguments

u

numeric matrix : one row per observation

copula

the copula object

...

other parameter to be passed to methods for this generic.

Value

The value of the copula on each observation

Functions

pCopula,matrix,ConvexCombCopula-method: Method for the cbCopula

pCopula,matrix,Cort-method: Method for the class Cort

pCopula,matrix,CortForest-method: Method for the class CortForest

pCopula,matrix,cbCopula-method: Method for the cbCopula

pCopula,matrix,cbkmCopula-method: Method for the cbCopula

Examples

cop <- cbCopula(cort::recoveryourself_data,m = 5)
pCopula(rep(0,2),cop) == 0
#> [1] TRUE
pCopula(rep(0.5,2),cop)
#> [1] 0.088
pCopula(rep(1,2),cop) == 1
#> [1] FALSE