Copula volume on hyper-boxes

u must be piecewise smaller than v, otherwise the function will return an error.

vCopula(u, v, copula, ...)

# S4 method for matrix,matrix
vCopula(u, v, copula)

Arguments

u

numeric matrix : minimum point of the hyper-rectangles, one row per observation.

v

numeric matrix : maximum point of the hyper-rectangle, one row per observation.

copula

the copula that we compute the measure on the box (u,v)

...

other parameter to be passed to methods for this generic.

Value

the measure of the copula.

Details

A method is currently implemented for the main virtual class 'Copula', but it assumes that a pCopula method is avaliable for the given copula. This method could be used with Copulas that are not from this package, assuming that pCopula(u,cop) works.

This function computes the measure of the copula according to the algorithm proposed by Cherubini U, Romagnoli S (2009-oct).

References

Cherubini U, Romagnoli S (2009). “Computing the Volume of n -Dimensional Copulas.” Applied Mathematical Finance, 16(4), 307--314. ISSN 1350-486X, 1466-4313.

Examples

cop <- cbCopula(LifeCycleSavings,m = 5)
vCopula(rep(0,5),rep(1,5),cop) == 1
#> [1] FALSE
vCopula(rep(0,5),rep(0.5,5),cop)
#> [1] 0.01875