The Cort estimator

Cort class

Cort(
  x,
  p_value_for_dim_red = 0.75,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose = TRUE
)

# S4 method for Cort
dim(x)

# S4 method for empiricalCopula
dim(x)

Arguments

x
The data, must be provided as a matrix with each row as an observation.

p_value_for_dim_red
a p_value for the localised dimension reduction test

min_node_size
The minimum number of observation avaliable in a leaf to initialise a split.

pseudo_data
set to True if you are already providing data on the copula space.

number_max_dim
The maximum number of dimension a split occurs in. Defaults to be all of the dimensions.

verbose
set to TRUE for more detailed output during the fit.

Value

a Cort object that can be fitted easily to produce a copula estimate.

Details

This class implements the CORT algorithm to a fit a multivariate copula using piece constant density.

Functions

dim,Cort-method: dimension of the copula

dim,empiricalCopula-method: dimension of the copula

Examples

library(cort)
(Cort(LifeCycleSavings[,1:3]))

#> Splitting...
#>      1 leaves to split...
#>            breakpoints : 0.4550217 0.5331809 0.490196 
#>            p_values    : 0.003003003 0 0 
#> 
#>      5 leaves to split...
#>            breakpoints : 0.2941176 0.3040362 0.9092953 
#>            p_values    : 0.7767768 0.2552553 0.1641642 
#> 
#>            breakpoints : 0.2011664 0.7173226 0.4313724 
#>            p_values    : 0.7937938 0.6166166 0.8558559 
#> 
#>            breakpoints : 0.9803922 0.3604974 0.4405684 
#>            p_values    : 0.04704705 0.1331331 0.1771772 
#> 
#>            breakpoints : 0.6496815 0.2156862 0.7118837 
#>            p_values    : 0.8388388 0.8248248 0.9269269 
#> 
#>            breakpoints : 0.9607844 0.7772614 0.1425547 
#>            p_values    : 0.2432432 0.6726727 0.6686687 
#> 
#>      4 leaves to split...
#>            breakpoints : 0.05882351 0.9673573 
#>            p_values    : 0.1151151 0.06306306 
#> 
#>            breakpoints : 0.5294118 0.5955174 
#>            p_values    : 0.5985986 0.4254254 
#> 
#>            breakpoints : 0.6671607 0.9803921 0.3375844 
#>            p_values    : 0.5425425 0.04004004 0.5545546 
#> 
#>            breakpoints : 0.7096229 0.7647059 0.07103174 
#>            p_values    : 0.5125125 0.05905906 0.1161161 
#> 
#>      1 leaves to split...
#>            breakpoints : 0.5294118 0.6923635 
#>            p_values    : 0.2422422 0.3573574 
#> 
#>      0 leaves to split...
#> Enforcing constraints...
#> -----------------------------------------------------------------
#>            OSQP v0.6.0  -  Operator Splitting QP Solver
#>               (c) Bartolomeo Stellato,  Goran Banjac
#>         University of Oxford  -  Stanford University 2019
#> -----------------------------------------------------------------
#> problem:  variables n = 49, constraints m = 98
#>           nnz(P) + nnz(A) = 1673
#> settings: linear system solver = qdldl,
#>           eps_abs = 1.0e-06, eps_rel = 1.0e-06,
#>           eps_prim_inf = 1.0e-06, eps_dual_inf = 1.0e-06,
#>           rho = 1.00e-01 (adaptive),
#>           sigma = 1.00e-06, alpha = 1.60, max_iter = 100000
#>           check_termination: on (interval 25),
#>           scaling: on, scaled_termination: off
#>           warm start: on, polish: on, time_limit: off
#> 
#> iter   objective    pri res    dua res    rho        time
#>    1  -1.2705e+05   3.04e-02   1.26e+07   1.00e-01   4.21e-04s
#>  200  -1.4186e+04   1.26e-03   2.58e+03   5.67e-01   1.69e-03s
#>  400  -5.3780e+02   4.44e-06   1.11e+02   3.81e-01   2.97e-03s
#>  600  -5.3216e+01   4.40e-07   1.11e+01   3.81e-01   4.14e-03s
#> 
#> status:               solved
#> solution polish:      unsuccessful
#> number of iterations: 600
#> optimal objective:    -53.2155
#> run time:             5.19e-03s
#> optimal rho estimate: 2.02e-01
#> 
#> Done !

#> Cort copula model: 50x3-dataset and 49 leaves