The Cort estimator

Cort class

Cort(
  data,
  p_value_for_dim_red = 0.75,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose = TRUE
)

Arguments

data
The data, must be provided as a matrix with each row as an observation.

p_value_for_dim_red
a p_value for the localised dimension reduction test

min_node_size
The minimum number of observation avaliable in a leaf to initialise a split.

pseudo_data
set to True if you are already providing data on the copula space.

number_max_dim
The maximum number of dimension a split occurs in. Defaults to be all of the dimensions.

verbose
set to TRUE for more detailed output during the fit.

Value

a Cort object that can be fitted easily to produce a copula estimate.

Details

This class implements the CORT algorithm to a fit a multivariate copula using piece constant density.

Examples

library(cort)
Cort(LifeCycleSavings)

#> Splitting...
#>      1 leaves to split...
#>            breakpoints : 0.4679535 0.4313724 0.4705358 0.3930684 0.5447669 
#>            p_values    : 0 0 0 0 0.01507538 
#> 
#>      10 leaves to split...
#>            breakpoints : 0.3344338 0.2992473 0.9476261 0.9607841 0.4726142 
#>            p_values    : 0.4773869 0.3015075 0.1708543 0.1105528 0.3768844 
#> 
#>            breakpoints : 0.1142149 0.6702751 0.3776094 0.1218912 0.01960776 
#>            p_values    : 1 0.6582915 0.3266332 0.3567839 0.160804 
#> 
#>            breakpoints : 0.2305007 0.9411766 0.1317509 0.1805018 0.7516609 
#>            p_values    : 0.6331658 0.1758794 0.04522613 0.4673367 0.5226131 
#> 
#>            breakpoints : 0.1439065 0.88433 0.3921569 0.5231292 0.3472491 
#>            p_values    : 0.2914573 0.2261307 0.6331658 1 0.5276382 
#> 
#>            breakpoints : 0.3725489 0.4704211 0.6391404 0.8717916 0.2417029 
#>            p_values    : 0.5175879 0 0.5577889 0.1356784 0.2361809 
#> 
#>            breakpoints : 0.6681163 0.3724812 0.7866212 0.8431373 0.407708 
#>            p_values    : 0.1356784 0.05527638 0.6331658 0.06030151 0.6080402 
#> 
#>            breakpoints : 0.7271346 0.1322973 0.7008493 0.6840725 0.6666667 
#>            p_values    : 0.1256281 0.8140704 0.2562814 0.2964824 0.4924623 
#> 
#>            breakpoints : 0.7331839 0.9102352 0.3229731 0.3921569 0.3963459 
#>            p_values    : 0.7638191 0.321608 1 0.8341709 1 
#> 
#>            breakpoints : 0.9607843 0.7839649 0.04412825 0.16251 0.8292015 
#>            p_values    : 1 1 0 0.1457286 1 
#> 
#>            breakpoints : 0.6364209 0.4705883 0.5332111 0.4904903 0.2087324 
#>            p_values    : 0.09045226 0 0.2361809 0 0.4020101 
#> 
#>      4 leaves to split...
#>            breakpoints : 0.7862366 0.235294 0.2690242 0.3511927 
#>            p_values    : 0.5879397 0.6180905 0.1306533 0.5678392 
#> 
#>            breakpoints : 0.7058823 0.595931 0.490952 0.7893786 
#>            p_values    : 0.07035176 0.07035176 0.2864322 0.5025126 
#> 
#>            breakpoints : 0.7254902 0.7711599 0.9128584 0.6002909 
#>            p_values    : 0.2562814 0.5728643 1 0.7135678 
#> 
#>            breakpoints : 0.9119542 0.9476338 0.8039576 0.745098 
#>            p_values    : 0.2361809 0.1005025 0.1256281 0.1256281 
#> 
#>      3 leaves to split...
#>            breakpoints : 0.7843137 0.3391889 0.3171114 0.3843885 
#>            p_values    : 0.1507538 0.3316583 0.5175879 0.2160804 
#> 
#>            breakpoints : 0.8845439 0.3153083 0.2426719 0.3039215 
#>            p_values    : 1 0.5678392 0.4221106 0.5376884 
#> 
#>            breakpoints : 0.6583994 0.8708779 0.6666667 
#>            p_values    : 0.120603 0.2462312 0.1809045 
#> 
#>      0 leaves to split...
#> Enforcing constraints...
#> -----------------------------------------------------------------
#>            OSQP v0.6.0  -  Operator Splitting QP Solver
#>               (c) Bartolomeo Stellato,  Goran Banjac
#>         University of Oxford  -  Stanford University 2019
#> -----------------------------------------------------------------
#> problem:  variables n = 317, constraints m = 458
#>           nnz(P) + nnz(A) = 29362
#> settings: linear system solver = qdldl,
#>           eps_abs = 1.0e-06, eps_rel = 1.0e-06,
#>           eps_prim_inf = 1.0e-06, eps_dual_inf = 1.0e-06,
#>           rho = 1.00e-01 (adaptive),
#>           sigma = 1.00e-06, alpha = 1.60, max_iter = 100000
#>           check_termination: on (interval 25),
#>           scaling: on, scaled_termination: off
#>           warm start: on, polish: on, time_limit: off
#> 
#> iter   objective    pri res    dua res    rho        time
#>    1  -1.4996e+02   6.64e-02   1.74e+05   1.00e-01   8.71e-03s
#>  200  -8.2113e+01   1.79e-04   7.36e-01   1.07e+00   3.29e-02s
#>  400  -7.9980e+01   1.45e-05   5.56e-02   1.07e+00   5.39e-02s
#>  600  -7.9844e+01   5.29e-06   7.78e-03   1.07e+00   7.49e-02s
#>  800  -7.9822e+01   5.54e-06   4.60e-03   1.07e+00   9.59e-02s
#> 1000  -7.9815e+01   5.53e-06   2.69e-03   1.07e+00   1.17e-01s
#> 1200  -7.9811e+01   5.51e-06   1.57e-03   1.07e+00   1.38e-01s
#> 1250  -7.9809e+01   4.06e-07   1.97e-03   1.07e+00   1.43e-01s
#> 
#> status:               solved
#> solution polish:      unsuccessful
#> number of iterations: 1250
#> optimal objective:    -79.8093
#> run time:             1.65e-01s
#> optimal rho estimate: 6.89e-01
#> 
#> Done !

#> Cort copula model: 50x5-dataset and 317 leaves